* Made maximum row length tests less strict * Moved microcircuit tests into test_connection_builder
