+++ tests/spinnaker/runtime/connection_builder/test_connection_builder.py
@@ -16,13 +16,13 @@ native_rng = sim.NativeRNG(NumpyRNG())
 # ----------------------------------------------------------------------------
 # Helpers
 # ----------------------------------------------------------------------------
-def _run_network(pop_size, connector, synapse, required_params, dt=1.0):
+def _run_network(pre_size, post_size, connector, synapse, required_params, dt=1.0):
     # Setup simulator
     sim.setup(timestep=dt, stop_after_loader=True)
 
     # Create two populations
-    pre = sim.Population(pop_size, sim.IF_curr_exp())
-    post = sim.Population(pop_size, sim.IF_curr_exp())
+    pre = sim.Population(pre_size, sim.IF_curr_exp())
+    post = sim.Population(post_size, sim.IF_curr_exp())
 
     # As we're not actually simulating the network, give a low
     # mean firing rate estimate so partitioning will be optimistic
@@ -61,7 +61,7 @@ def _run_network(pop_size, connector, synapse, required_params, dt=1.0):
                               weight=0.0, delay=10.0)])
 def test_static_delay(pop_size, synapse):
     # Run network
-    proj, proj_data = _run_network(pop_size, sim.AllToAllConnector(),
+    proj, proj_data = _run_network(pop_size, pop_size, sim.AllToAllConnector(),
                                    synapse, ["delay"])
 
     # Check they all match
@@ -74,7 +74,7 @@ def test_static_delay(pop_size, synapse):
                           sim.FixedProbabilityConnector(0.1, rng=native_rng)])
 def test_connector_metrics(pop_size, connector):
     # Run network
-    proj, proj_data = _run_network(pop_size, connector,
+    proj, proj_data = _run_network(pop_size, pop_size, connector,
                                    sim.StaticSynapse(weight=0.0, delay=1.0),
                                    [])
 
@@ -100,7 +100,7 @@ def test_connector_metrics(pop_size, connector):
     assert estimated_max_row_synapses >= actual_max_row_synapses
 
     # Check estimated maximum isn't TOO big
-    assert estimated_max_row_synapses <= actual_max_row_synapses * 1.25
+    assert estimated_max_row_synapses <= actual_max_row_synapses * 3
 
     # Check estimated mean is approximately correct
     assert estimated_mean_row_synapses <= (1.25 * actual_mean_row_synapses)
@@ -116,7 +116,9 @@ def test_connector_metrics(pop_size, connector):
                           (200, 0.2, True),
                           (200, 0.2, False),
                           (2000, 0.05, True),
-                          (2000, 0.2, True)])
+                          pytest.mark.xfail(reason="Issue #65")((2000, 0.05, False)),
+                          (2000, 0.2, True),
+                          pytest.mark.xfail(reason="Issue #65")((2000, 0.2, False))])
 def test_fixed_number_connector(pop_size, connection_proportion, with_replacement):
 
     num_connections = int(connection_proportion * pop_size**2)
@@ -124,7 +126,7 @@ def test_fixed_number_connector(pop_size, connection_proportion, with_replacemen
                                               with_replacement=with_replacement,
                                               rng=native_rng)
 
-    proj, proj_data = _run_network(pop_size, connector,
+    proj, proj_data = _run_network(pop_size, pop_size, connector,
                                    sim.StaticSynapse(weight=0.0, delay=1.0),
                                    [])
 
@@ -166,7 +168,7 @@ def test_connector_delay_dist(delay_dist_name_params):
                                     delay_dist_name_params[1],
                                     rng=native_rng)
     # Run network
-    proj, proj_data = _run_network(1000, sim.AllToAllConnector(),
+    proj, proj_data = _run_network(1000, 1000, sim.AllToAllConnector(),
                                    sim.StaticSynapse(weight=0.0, delay=delay_dist),
                                    ["delay"], 0.1)
 
@@ -189,7 +191,7 @@ def test_connector_weight_dist(weight_dist_name_params):
                                      rng=native_rng)
 
     # Run network
-    proj, proj_data = _run_network(1000, sim.AllToAllConnector(),
+    proj, proj_data = _run_network(1000, 100, sim.AllToAllConnector(),
                                    sim.StaticSynapse(weight=weight_dist, delay=1.0),
                                    ["weight"], 1.0)
 
@@ -203,3 +205,59 @@ def test_connector_weight_dist(weight_dist_name_params):
     # Run Kolmogorov-Smirnov test on data
     test = ks_2samp(proj_data[2], samples)
     assert test[1] > 0.05
+
+@pytest.mark.parametrize("pop_size_1,pop_size_2,num_connections",
+                         [(547,547,52147),(106,1439,3049),(294,583,0),(294,294,13506),(485,547,877),
+                          (294,2068,0),(1439,2068,22711),(485,106,3181),(485,485,20407),(106,294,251),
+                          (547,1439,13198),(2191,583,41025),(1439,106,1318),(583,485,18171),(583,2068,223203),
+                          (583,106,1686),(583,583,50153),(294,106,0),(294,2191,0),(294,1439,108239),
+                          (2191,2068,202507),(2068,2191,35029),(547,294,81),(2068,547,81009),(2191,2191,244773),
+                          (2068,294,22547),(106,106,4284),(106,2068,0),(2068,1439,46796),(2191,1439,67252),
+                          (106,485,24079),(106,583,0),(2191,547,99172),(547,106,128),(583,294,172),
+                          (2068,2068,454932),(1439,485,14390),(1439,547,87964),(1439,1439,83697),(2068,485,106136),
+                          (1439,2191,146211),(1439,294,28806),(485,2191,7144),(2068,583,174317),(547,485,1519),
+                          (2191,294,2194),(547,2068,96695),(485,2068,32931),(547,2191,174096),(583,2191,7564),
+                          (485,1439,41108),(583,547,927),(485,294,4005),(547,583,16889),(1439,583,3532),
+                          (294,547,0),(294,485,0),(583,1439,5559),(2068,106,12356),(485,583,22197),
+                          (2191,485,55078),(106,547,0),(106,2191,70),(2191,106,6048)])
+def test_microcircuit_projections(pop_size_1, pop_size_2, num_connections):
+
+    connector = sim.FixedTotalNumberConnector(num_connections,
+                                              with_replacement=True,
+                                              rng=native_rng)
+
+    # Estimate maximum and mean row length
+    pre_slice = UnitStrideSlice(0, pop_size_1)
+    post_slice = UnitStrideSlice(0, pop_size_2)
+
+    # Distribution over synapses per row per post slice
+    row_synapses_distribution = connector._row_synapses_distribution(
+        pre_slice, post_slice, pop_size_1, pop_size_2)
+
+    estimated_max_row_synapses = row_synapses_distribution.ppf(0.9999 ** (1.0 / pop_size_1))
+    estimated_mean_row_synapses = row_synapses_distribution.mean()
+
+    proj, proj_data = _run_network(pop_size_1, pop_size_2, connector,
+                                   sim.StaticSynapse(weight=0.0, delay=0.1),
+                                   [], 0.1)
+
+    # Build row-length histogram
+    row_length_histogram = binned_statistic(proj_data[0], proj_data[1],
+                                            statistic="count", bins=range(pop_size_2 + 1))[0]
+
+    # Test that the number of connections is correct
+    assert int(row_length_histogram.sum()) == num_connections
+
+    # Computer max and mean
+    actual_max_row_synapses = np.amax(row_length_histogram)
+    actual_mean_row_synapses = np.average(row_length_histogram)
+
+     # Check estimated maximum is greater or equal than actual maximum
+    assert estimated_max_row_synapses >= actual_max_row_synapses
+
+    # Check estimated maximum isn't TOO big
+    assert estimated_max_row_synapses <= actual_max_row_synapses * 3.0
+
+    # Check estimated mean is approximately correct
+    assert estimated_mean_row_synapses <= (1.25 * actual_mean_row_synapses)
+    assert estimated_mean_row_synapses >= (0.75 * actual_mean_row_synapses)
